implement getModificationTimestamps with custom element filter
this makes the non-old-style-multipolygon path in CellIterator.iterateAll()
much faster, because not it can preemtively filter out element that
don't need to be looked at and for the rest only calculate the
geometries at the timestamps where something actually changed.

The effect is largest if a fine timegrid is chosen and the objects
don't change often (e.g. for buildings in Heidelberg at a monthly
time grid, only 1/30th of the geometry operations need to be
performed, for highways it's ~1/20th). Total speedup is not as
large for "coarse" (e.g. monthly) time grids other fix-costs (e.g.
parsing) dominate (more like 30% total speedup). But for very
fine time intervals (e.g. daily) it makes a huge difference in
both processing time and memory consumption.