implement flatMapReduce/Aggregate (and *GroupedById) functions
flatMap functions are used when there can be multiple results in a single call of the map function (0 to many, e.g. when a linestring is in multiple osmatrix cells)

also available: a map/reducer which gets by osm-id grouped lists of what the mapper function normally gets, and expects another list as output (doesn't have to be a n:n relationship, the function can return an arbitrary number of results).

this can be used to calculate metrics that depend on the full history of a single osm entity, e.g. it could be used to identify edit wars, to search for complex editing patterns, etc.

still missing: full implementation in the Ignite endpoint implementation, partly missing implementation of the multithreaded H2