make mapReducer/Aggregator objects free of side effects
Instead of altering the internal state of mapReducer/mapAggregator
objects, all* functions like `map`, `where`, etc. return new objects
that are a copy of the original mapReducer/Aggreagtor with the
updated internal state.

This allows re-use of existing mapReducer and mapAggregator objects
and prevents issues like #65. closes #65

* the only functions that now don't return a copied version of the
original object are the `reduce` (and it's shorthands like `sum`,
`collect`, etc.) and some of the settings like `keytables` or
`zerofill` (of MapAggregatorByTimestamp).