mark keytables dependent attributes of MapReducer transient
this is necessary to deploy them seamlessly on ignite.

Keytables are only used in the definition phase of the MapReducer (setting tag filters, initializing the TagInterpreter), but not during the actual parallel execution of the reduce step on the cluster, therefore these properties don't need to be serialized for running it on ignite. This is fine https://i.imgur.com/c4jt321.png