implement most simple ignite backend: affinityCall implementation
It's just sending separate affinityCalls to the cluster and reducing all cell's results locally on the client.

There for testing purposes and possibly for circumstances where one knows that only few cells have to be iterated over (e.g. queries in a small area of interest), where the (~constant) overhead associated with the other methods might be larger than the (~linear) inefficiency with this implementation