implement much faster bbox-in-polygon test (for CellIterator)
This implements the idea from https://blog.jochentopf.com/2017-02-06-expedicious-and-exact-extracts-with-osmium.html (slightly enhanced to also handle the bbox-in-polygon case) for major speed ups in the CellIterator and oshdb-api when working with complex bounding polygons (areas of interest).

In a quick test with the boundary of Birmingham (~3000 points) the speed-up was more than x100 (!) for the actual bbox-in-polygon test (resulting in an overall speed-up of x20, but that depends largely on the actual analysis one's performing).