Revert "add "getAggregtionTimestampInteval" to OSMContribution objects"
This reverts commit 3680ff2c8f0ab465cc92aa8dc6b574c5da17c51d.

this property has never actually used in merge request !28, since by-timestamp-aggregation logic was implemented differently in the end.

PS: it could maybe still be interesting to think about using this kind of method to get more meaningful results for timestamp-aggregated contribution results (e.g. by doing `aggregateBy(OSMContribution:getAggregationTimestampInterval)` instead of simply `aggregateByTimestamp()`) to get more meaningful keys in the result object. Downside of this is that it's longer to write, harder to remember and that zerofilling wouldn't work out of the box.