fix crash when trying to build geometries with broken node refs
example:

way 177974941 in version 6 (2015-05-15T14:54:14Z) references node 3516928189. This node's version 1 is redacted, thus only starts to "exist" from version 2 (2015-06-08T18:50:54Z).
When requesting the way's geometry between these dates, the call to way.getRefEntities returns a null value, which need to be filtered to not cause a crash.

see:
  https://www.openstreetmap.org/api/0.6/way/177974941/history
  https://www.openstreetmap.org/api/0.6/node/3516928189/history