drop `Filter.by` helpers
Reason: They did not really fulfil a specific purpose because one can already "and" combine OSH+OSM filters by calling `.filter` twice, and the geometry callback filter basically just replicates the `.filter(Predicate)` anyway.
Also the given documentation and examples were not in line with the actual behaviour when one would have negated the filter, so it in the end just confusing to use.